refactor auth

+++ api/app/db/queries/auth.ts
@@ -4,8 +4,8 @@ import {JWT_REFRESH_TTL} from '../../services/auth'
 function storeRefreshToken(data) {
     return knex('refresh_tokens').insert(data)
 }
-function removeRefreshToken(id: string): Promise<number> {
-    return knex('refresh_tokens').where({id}).delete()
+function removeRefreshToken(id: number): Promise<number> {
+    return knex('refresh_tokens').where({user_id: id}).delete()
 }
 function checkRefreshToken(id: string): Promise<{id: string, user_id: string}> {
     if (!id) {
